perf: modernize and improve example doc structure #105 #107 example code updated to reflect latest Auro element APIs example documentation layout improved custom registration updated and moved to install page core element descriptions updated properties now listed alphabetically doc build system now uses the latest version of auro-cli package json configuration updated to support latest auro-cli additional packages dependencies updated to latest versions
