+++ .gitignore
@@ -24,3 +24,5 @@ coverage/
 ## Added by Auro CLI migration (github-config-and-gitignore-changes)
 demo/*.md
 demo/*.min.js
+
+custom-elements.json
