+++ apiExamples/basic.html
@@ -1,10 +1,14 @@
-<auro-carousel centerSelected>
+<auro-carousel>
   <auro-pane date="2020-10-10" price="$435"></auro-pane>
   <auro-pane date="2020-10-11" price="$435"></auro-pane>
   <auro-pane date="2020-10-12" price="$435"></auro-pane>
-  <auro-pane date="2020-10-13" selected price="$666"></auro-pane>
-  <auro-pane date="2020-10-14" price="$435" disabled></auro-pane>
+  <auro-pane date="2020-10-13" price="$435"></auro-pane>
+  <auro-pane date="2020-10-14" price="$435"></auro-pane>
   <auro-pane date="2020-10-15" price="$435"></auro-pane>
+  <auro-pane date="2020-10-16" price="$435"></auro-pane>
+  <auro-pane date="2020-10-17" price="$435"></auro-pane>
+  <auro-pane date="2020-10-18" selected price="$435"></auro-pane>
+  <auro-pane date="2020-10-19" price="$435"></auro-pane>
   <auro-pane date="2020-10-16" price="$435"></auro-pane>
   <auro-pane date="2020-10-17" price="$435"></auro-pane>
   <auro-pane date="2020-10-18" price="$435"></auro-pane>
