--- apiExamples/default.html
@@ -1,16 +1,0 @@
-<auro-carousel>
-  <auro-pane date="2020-10-10" price="$435"></auro-pane>
-  <auro-pane date="2020-10-11" price="$435"></auro-pane>
-  <auro-pane date="2020-10-12" price="$435"></auro-pane>
-  <auro-pane date="2020-10-13" selected price="$435"></auro-pane>
-  <auro-pane date="2020-10-14" price="$435" disabled></auro-pane>
-  <auro-pane date="2020-10-15" price="$435"></auro-pane>
-  <auro-pane date="2020-10-16" price="$435"></auro-pane>
-  <auro-pane date="2020-10-17" price="$435"></auro-pane>
-  <auro-pane date="2020-10-18" price="$435"></auro-pane>
-  <auro-pane date="2020-10-19" price="$435"></auro-pane>
-  <auro-pane date="2020-10-16" price="$435"></auro-pane>
-  <auro-pane date="2020-10-17" price="$435"></auro-pane>
-  <auro-pane date="2020-10-18" price="$435"></auro-pane>
-  <auro-pane date="2020-10-19" price="$435"></auro-pane>
-</auro-carousel>
