+++ docs/api.md
@@ -1,43 +1,43 @@
 # auro-carousel
 
-The auro-carousel component displays a group of elements in a scrollable container.
+The `auro-carousel` element displays a group of elements in a scrollable container.
 
-## Properties
+### Properties & Attributes
 
-| Property         | Attribute        | Type      | Default | Description                                      |
-|------------------|------------------|-----------|---------|--------------------------------------------------|
-| `centerSelected` | `centerSelected` | `String`  |         | Apply to outer auro-carousel element to automatically center the selected node on UI render. |
-| `displayArrows`  | `displayArrows`  | `Boolean` |         | Forces left and right navigation to stick in DOM regardless of content width |
-| `label`          | `label`          | `String`  |         | DEPRECATED - Use `ariaLabel` slot instead.       |
-| `scrollDistance` | `scrollDistance` | `Number`  | 300     | How many pixels to scroll the carousel when the shoulder buttons are triggered. |
+| Properties     | Attributes     | Modifiers | Type    | Default | Description                                                                                    |
+| -------------- | -------------- | --------- | ------- | ------- | ---------------------------------------------------------------------------------------------- |
+| centerSelected | centerSelected |           | string  |         | Automatically centers the selected node on UI render                                           |
+| displayArrows  | displayArrows  |           | boolean |         | Always displays the left and right arrows, regardless of content width                         |
+| label          | label          |           | string  |         | DEPRECATED - Use `ariaLabel` slot instead.                                                     |
+| scrollDistance | scrollDistance |           | number  | `300`   | The number of pixels to scroll the carousel when the left or right buttons are interacted with |
 
-## Methods
+### Methods
 
-| Method                  | Type                  | Description                                      |
-|-------------------------|-----------------------|--------------------------------------------------|
-| `actionOnChildrenReady` | `(): void`            | Function handler for anything that happens when all its children is ready. |
-| `centerElement`         | `(el: Element): void` | Centers the given element inside the carousel.<br /><br />**el**: The element to center inside the carousel. Must be a descendant of the carousel. |
-| `scrollCarousel`        | `(num: number): void` | Scrolls the carousel by the given amount.<br /><br />**num**: The number of pixels to scroll the carousel by. Positive scrolls to the<br />right, negative scrolls to the left. |
-| `scrollToSelected`      | `(): void`            | Scroll to the first child component that have 'selected' attribute. |
+| Name             | Parameters                                                                                                                       | Return          | Description                                                         |
+| ---------------- | -------------------------------------------------------------------------------------------------------------------------------- | --------------- | ------------------------------------------------------------------- |
+| centerElement    | `el` (Element) - The element to center inside the carousel. Must be a descendant of the carousel.                                | [object Object] | Centers the given element inside the carousel.                      |
+| register         | `name` (string) - The name of the element that you want to register.                                                             |                 | This will register this element with the browser.                   |
+| scrollCarousel   | `num` (number) - The number of pixels to scroll the carousel by. Positive scrolls to the<br>right, negative scrolls to the left. | [object Object] | Scrolls the carousel by the given amount.                           |
+| scrollToSelected | None                                                                                                                             | [object Object] | Scroll to the first child component that have 'selected' attribute. |
 
-## Events
+### Events
 
-| Event         | Type               | Description                                      |
-|---------------|--------------------|--------------------------------------------------|
-| `scrollLeft`  | `CustomEvent<any>` | when the guest clicks the 'left' carousel button |
-| `scrollRight` | `CustomEvent<any>` | when the guest clicks the 'right' carousel button |
+| Name        | Description                                                  |
+| ----------- | ------------------------------------------------------------ |
+| scrollLeft  | Dispatched when the guest clicks the 'left' carousel button  |
+| scrollRight | Dispatched when the guest clicks the 'right' carousel button |
 
-## Slots
+### Slots
 
-| Name                     | Description                                      |
-|--------------------------|--------------------------------------------------|
-|                          | the elements in the carousel                     |
-| `ariaLabel`              | Text to give an accessible name to the carousel. |
-| `ariaLabel.scroll.left`  | Text to give an accessible name to the left scroll button. |
-| `ariaLabel.scroll.right` | Text to give an accessible name to the right scroll button. |
+| Name                   | Description                                                |
+| ---------------------- | ---------------------------------------------------------- |
+| (default)              | Default slot for the elements in the carousel              |
+| ariaLabel              | Text to give an accessible name to the carousel            |
+| ariaLabel.scroll.left  | Text to give an accessible name to the left scroll button  |
+| ariaLabel.scroll.right | Text to give an accessible name to the right scroll button |
 
-## CSS Shadow Parts
+### CSS Shadow Parts
 
-| Part      | Description                                     |
-|-----------|-------------------------------------------------|
-| `wrapper` | The primary root HTML element of the component. |
+| Name    | Description                                     |
+| ------- | ----------------------------------------------- |
+| wrapper | The primary root HTML element of the component. |
