+++ docs/partials/customRegistration.md
@@ -1,1 +1,9 @@
-<!-- add custom registration content here -->
+```js
+// Import the class only
+import { AuroCarousel } from '@aurodesignsystem/auro-carousel/class';
+
+// Register with a custom name if desired
+AuroCarousel.register('custom-carousel');
+```
+
+This will create a new custom element `<custom-carousel>` that behaves exactly like `<auro-carousel>`, allowing both to coexist on the same page without interfering with each other.
